Prevent accidental loop
superagent itself defines `Request.prototype._end()` and expects to be
able to call it (from `.end()`) to kick off an actual request.
Therefore, `_end` is a name we **should not** use internally, especially
as one that ends up calling the underlying `.end()`

Instead, use a semi-namespaced key to store off the existing value so
that it's safe even if another plugin does something similar.